OPENAI Project/requirements.txt
python-docx
openai>=1.0.0
python-dotenv